fix: keep large trade values and quantities in long arithmetic

+++ src/main/java/com/flippingutilities/model/FlippingItem.java
@@ -242,7 +242,7 @@ public class FlippingItem implements Searchable
 		return HistoryManager.getTotalRevenueOrExpense(tradeList, isBuy);
 	}
 
-	public static int countFlipQuantity(List<OfferEvent> tradeList)
+	public static long countFlipQuantity(List<OfferEvent> tradeList)
 	{
 		return HistoryManager.countFlipQuantity(tradeList);
 	}
